backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
pydantic==2.6.4
tenacity==8.2.3
openai==1.14.3
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
beautifulsoup4==4.12.3
playwright==1.43.0
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
google-generativeai==0.4.1
psycopg2-binary==2.9.9
